server/requirements.txt
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorboard==2.10.1
pillow==8.4.0
protobuf==3.19.6
h5py==3.1.0
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
tensorflow-io-gcs-filesystem==0.31.0
bchlib==0.10
numpy>=1.21.0

